// script-generator/src/constants/api.ts
export const Locations = {
  DECK: 0x01,
  HAND: 0x02,
  MZONE: 0x04,
  SZONE: 0x08,
  GRAVE: 0x10,
  REMOVED: 0x20,
  EXTRA: 0x40,
  OVERLAY: 0x80,
  ONFIELD: 0x0C,
  FZONE: 0x100,
  PZONE: 0x200,
  DECKBOT: 0x10001,
  DECKSHF: 0x20001,
} as const;

// Positions
export const Positions = {
  FACEUP_ATTACK: 0x1,
  FACEDOWN_ATTACK: 0x2,
  FACEUP_DEFENSE: 0x4,
  FACEDOWN_DEFENSE: 0x8,
  FACEUP: 0x5,
  FACEDOWN: 0xA,
  ATTACK: 0x3,
  DEFENSE: 0xC,
} as const;

// Types
export const Types = {
  MONSTER: 0x1,
  SPELL: 0x2,
  TRAP: 0x4,
  NORMAL: 0x10,
  EFFECT: 0x20,
  FUSION: 0x40,
  RITUAL: 0x80,
  TRAPMONSTER: 0x100,
  SPIRIT: 0x200,
  UNION: 0x400,
  DUAL: 0x800,
  TUNER: 0x1000,
  SYNCHRO: 0x2000,
  TOKEN: 0x4000,
  QUICKPLAY: 0x10000,
  CONTINUOUS: 0x20000,
  EQUIP: 0x40000,
  FIELD: 0x80000,
  COUNTER: 0x100000,
  FLIP: 0x200000,
  TOON: 0x400000,
  XYZ: 0x800000,
  PENDULUM: 0x1000000,
  SPSUMMON: 0x2000000,
  LINK: 0x4000000,
} as const;

// Attributes
export const Attributes = {
  EARTH: 0x01,
  WATER: 0x02,
  FIRE: 0x04,
  WIND: 0x08,
  LIGHT: 0x10,
  DARK: 0x20,
  DEVINE: 0x40,
} as const;

// Races
export const Races = {
  WARRIOR: 0x1,
  SPELLCASTER: 0x2,
  FAIRY: 0x4,
  FIEND: 0x8,
  ZOMBIE: 0x10,
  MACHINE: 0x20,
  AQUA: 0x40,
  PYRO: 0x80,
  ROCK: 0x100,
  WINDBEAST: 0x200,
  PLANT: 0x400,
  INSECT: 0x800,
  THUNDER: 0x1000,
  DRAGON: 0x2000,
  BEAST: 0x4000,
  BEASTWARRIOR: 0x8000,
  DINOSAUR: 0x10000,
  FISH: 0x20000,
  SEASERPENT: 0x40000,
  REPTILE: 0x80000,
  PSYCHO: 0x100000,
  DEVINE_RACE: 0x200000,
  CREATORGOD: 0x400000,
  WYRM: 0x800000,
  CYBERSE: 0x1000000,
  ILLUSION: 0x2000000,
} as const;

// Reason
export const Reasons = {
  DESTROY: 0x1,
  RELEASE: 0x2,
  TEMPORARY: 0x4,
  MATERIAL: 0x8,
  SUMMON: 0x10,
  BATTLE: 0x20,
  EFFECT: 0x40,
  COST: 0x80,
  ADJUST: 0x100,
  LOST_TARGET: 0x200,
  RULE: 0x400,
  SPSUMMON: 0x800,
  DISSUMMON: 0x1000,
  FLIP: 0x2000,
  DISCARD: 0x4000,
  RDAMAGE: 0x8000,
  RRECOVER: 0x10000,
  RETURN: 0x20000,
  FUSION: 0x40000,
  SYNCHRO: 0x80000,
  RITUAL: 0x100000,
  XYZ: 0x200000,
  REPLACE: 0x1000000,
  DRAW: 0x2000000,
  REDIRECT: 0x4000000,
  REVEAL: 0x8000000,
  LINK: 0x10000000,
  LOST_OVERLAY: 0x20000000,
  MAINTENANCE: 0x40000000,
  ACTION: 0x80000000,
} as const;

// Phase
export const Phases = {
  DRAW: 0x01,
  STANDBY: 0x02,
  MAIN1: 0x04,
  BATTLE_START: 0x08,
  BATTLE_STEP: 0x10,
  DAMAGE: 0x20,
  DAMAGE_CAL: 0x40,
  BATTLE: 0x80,
  MAIN2: 0x100,
  END: 0x200,
} as const;

// Link Markers
export const LinkMarkers = {
  BOTTOM_LEFT: 0x001,
  BOTTOM: 0x002,
  BOTTOM_RIGHT: 0x004,
  LEFT: 0x008,
  RIGHT: 0x020,
  TOP_LEFT: 0x040,
  TOP: 0x080,
  TOP_RIGHT: 0x100,
} as const;

// Categories (for effect hints)
export const Categories = {
  DESTROY: 0x1,
  RELEASE: 0x2,
  TOHAND: 0x4,
  TODECK: 0x8,
  TOGRAVE: 0x10,
  REMOVE: 0x20,
  SPECIAL_SUMMON: 0x40,
  DRAW: 0x80,
  DAMAGE: 0x100,
  RECOVER: 0x200,
  DECKDES: 0x400,
  HANDDES: 0x800,
  SUMMON: 0x1000,
  FLIP: 0x2000,
  POSITION: 0x4000,
  CONTROL: 0x8000,
  DISABLE: 0x10000,
  ATKCHANGE: 0x20000,
  DEFCHANGE: 0x40000,
  COUNTER: 0x80000,
  EQUIP: 0x100000,
  FLAG: 0x200000,
  ATKNDEF: 0x400000,
  HINT: 0x800000,
  SEARCH: 0x1000000,
  GRAVE_ACTION: 0x2000000,
  COIN: 0x4000000,
  DICE: 0x8000000,
} as const;

// Summon Types
export const SummonTypes = {
  NORMAL: 0x10000000,
  ADVANCE: 0x11000000,
  SPECIAL: 0x20000000,
  FUSION: 0x21000000,
  RITUAL: 0x22000000,
  SYNCHRO: 0x23000000,
  XYZ: 0x24000000,
  PENDULUM: 0x25000000,
  LINK: 0x26000000,
} as const;

// Effect Types
export const EffectTypes = {
  SINGLE: 0x0001,
  FIELD: 0x0002,
  EQUIP: 0x0004,
  TARGET: 0x0008,
  CONTINUOUS: 0x0010,
  QUICK_O: 0x0020,
  QUICK: 0x0040,
  TRIGGER_O: 0x0080,
  TRIGGER_F: 0x0100,
  TRIGGER: 0x0200,
  IGNITION: 0x0400,
  ACTION: 0x0800,
  FLIP: 0x1000,
  CONTINUOUS_O: 0x2000,
  CONTINUOUS_F: 0x4000,
  ALL: 0x7F7F,
} as const;
// ...
export const Events = {
  STARTUP: 1000,
  FLIP: 1001,
  FREE_CHAIN: 1002,
  STANDBY: 1003,
  DESTROY: 1010,
  MOVE: 1011,
  LEAVE_FIELD: 1012,
  CHANGE_POS: 1013,
  RELEASE: 1014,
  DISCARD: 1015,
  TO_GRAVE: 1016,
  RETURN_TO_GRAVE: 1017,
  TO_HAND: 1018,
  RETURN_TO_HAND: 1019,
  TO_DECK: 1020,
  RETURN_TO_DECK: 1021,
  REMOVE: 1022,
  TO_EXCLUDED: 1023,
  CONTROL_CHANGED: 1024,
  SUMMON_SUCCESS: 1025,
  FLIP_SUMMON_SUCCESS: 1026,
  SPSUMMON_SUCCESS: 1027,
  MSET_SUCCESS: 1028,
  SSET_SUCCESS: 1029,
  SUMMON: 1030,
  FLIP_SUMMON: 1031,
  SPSUMMON: 1032,
  MSET: 1033,
  SSET: 1034,
  BE_PRE_MATERIAL: 1035,
  BE_PRE_TARGET: 1036,
  BE_MATERIAL: 1037,
  BE_CHAIN_TARGET: 1038,
  BE_BATTLE_TARGET: 1039,
  BATTLE_START: 1040,
  BATTLE_CONFIRM: 1041,
  BATTLE_END: 1042,
  BATTLED: 1043,
  BATTLE_DESTROYING: 1044,
  BATTLE_DESTROYED: 1045,
  PRE_BATTLE_DAMAGE: 1046,
  BATTLE_DAMAGE: 1047,
  DAMAGE: 1048,
  RECOVER: 1049,
  PAY_LPCOST: 1050,
  ADD_COUNTER: 1051,
  REMOVE_COUNTER: 1052,
  REPLACE_COUNTER: 1053,
  CHAIN_END: 1054,
  CHAIN_SOLVING: 1055,
  CHAIN_NEGATED: 1056,
  CHAIN_DISABLED: 1057,
  ADJUST: 1058,
  RESETS_CANCEL: 1059,
  TURN_END: 1060,
  PRE_DRAW: 1061,
  DRAW: 1062,
  LEVEL_UP: 1063,
  BE_SPECIAL_SUMMONED: 1064,
  END_PHASE: 1065,
  TO_DGRAVE: 1066,
  TO_DHAND: 1067,
  TO_DDECK: 1068,
  PRE_DESTROY: 1069,
  PRE_MOVE: 1070,
  PRE_RELEASE: 1071,
  PRE_TO_GRAVE: 1072,
  PRE_TO_HAND: 1073,
  PRE_TO_DECK: 1074,
  PRE_REMOVE: 1075,
  PRE_SUMMON: 1076,
  PRE_FLIP_SUMMON: 1077,
  PRE_SPSUMMON: 1078,
  PRE_MSET: 1079,
  PRE_SSET: 1080,
  PRE_DAMAGE: 1081,
  PRE_RECOVER: 1082,
  PRE_PAY_LPCOST: 1083,
  PRE_ADD_COUNTER: 1084,
  PRE_REMOVE_COUNTER: 1085,
  PRE_ATTACK: 1086,
  PRE_BATTLE_DAMAGE_STEP: 1087,
  CUSTOM: 1099,
} as const;